requirements.txt
mouseinfo==0.1.3
numpy==2.2.6
opencv-python==4.13.0.92
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pillow==12.1.1
pip @ file:///home/task_177071101768311/croot/pip_1770711076895/work
pyautogui==0.9.54
pygetwindow==0.0.9
pymsgbox==2.0.1
pynput==1.8.1
pyperclip==1.11.0
pyrect==0.2.0
pyscreeze==1.0.1
pytweening==1.2.0
setuptools==80.10.2
six==1.17.0
wheel==0.46.3
